Game renders to texture!
I think I have this figured out. The screen dimensions are small
(256x240), and the window dimensions are stored in the WSL_App struct.
When it comes time to render, things are drawn in the correct place on
the small resolution screen, then the screen is scaled to fit the window
dimensions, then drawn centered in the window.