Basic physics/movement!
Lots of work on the rendering pipeline and the main loop. The graphics
arent as "crisp" as I want them to be, and I'm totally stumped with what
I could be doing differently. Switched the main loop from a variable
time step to a fixed time step. I need to continue working on
synchronizing the update/render routines?

Added in some basic physics for gravity and friction. The player entity
(which is really the only entity right now) moves with a good "feel",
accelerating to full speed the longer you hold the arrow keys. Also, the
player doesn't "stop" immediately when moving, but takes a step after
you release the key (friction). Jumping could be improved - but it's
pretty good now.

Roughed in the ideas for animation, not sure exactly how I want to
handle animated sprites BUT the ideas are there. The player's animation
for movement/jumping looks pretty good (except it blurs going across the
screen - not "crisp").