More tile work!
Platforms (sorta) work - I need to spend some time thinking about
collision handling, and how I actually want the map stored. It probably
would make more sense to have entities and map tiles be in separate
lists - but for now this is sufficient.

Tiles can be flagged as animated, and then given a sprite to updte to
with a two-frame animation. Kinda hacked in there for now, just to see
it working. I included a couple example xml/tmx files in assets to play
around with, and they can be launched by "./1BP mapfile.xml"