Still working on collision detection...
Keep doing the one step forward/two steps back with this. Clearly, this
is **the** most critical part of a platformer style game - so it needs
to be SOLID before moving on to the next challenge. I moved the code for
resolving the movement to it's own function, and have been playing with
checking each entity for x/y collisions in one loop (x before y). I
think this is the most solid this has felt so far - but still has a ways
to go.