More collision work
Hitboxes for platforms are now adjusted so that they are *slightly*
smaller than the tile size on sides that aren't adjacent to a
tile/platform - which makes it so the player can actually fall off the
tile. Might take a break from doing physics/collision tweaking and start
working on scrolling maps or something fun.